requirements_gradio.txt
torch==2.1.0
torchvision==0.16.0
gradio==3.50.2
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
